backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
email-validator>=2.0.0
alembic==1.12.1
pydantic-settings==2.1.0
redis==5.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
prometheus-client>=0.19.0
python-json-logger>=2.0.7
pytest==7.4.4
httpx==0.26.0
openai>=1.0.0
anthropic>=0.21.0
